Refactor for OOP
Minor rafactor of development branch
