+++ app/src/main/java/com/example/divvy/Controllers/NavBarController.java
@@ -3,9 +3,8 @@ package com.example.divvy.Controllers;
 import android.content.Context;
 import android.content.Intent;
 import android.support.design.widget.BottomNavigationView;
-import android.support.annotation.NonNull;
-import android.view.MenuItem;
 
+import com.example.divvy.Controllers.helpers.LoginAuthenticator;
 import com.example.divvy.R;
 
 public class NavBarController {
@@ -15,7 +14,7 @@ public class NavBarController {
             Intent intent = null;
             switch (item.getItemId()) {
                 case R.id.navigation_home:
-                    intent = new Intent(context, UserProfileActivity.class);
+                    intent = new Intent(context, UserProfileViewController.class);
                     intent.putExtra("owner", LoginAuthenticator.getInstance().getUser(context));
                     break;
                 case R.id.navigation_dashboard:
