+++ app/src/main/java/com/example/divvy/Controllers/Services/RegService.java
@@ -1,24 +1,13 @@
-package com.example.divvy;
+package com.example.divvy.Controllers.Services;
 
 import android.app.IntentService;
 import android.content.Context;
 import android.content.Intent;
-import android.net.Network;
 import android.os.Bundle;
 import android.os.ResultReceiver;
 import android.support.annotation.Nullable;
 
-import com.example.divvy.models.Listing;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-
-import static com.example.divvy.httprequest.get;
 
 public class RegService extends IntentService {
     public RegService() {
